remove optionals from pipeline and stagedefinition, make tests use yaml instead of toml

+++ api/src/main/java/de/itdesigners/winslow/api/pipeline/ResourceInfo.java
@@ -5,12 +5,12 @@ import java.beans.ConstructorProperties;
 
 public class ResourceInfo {
     public final           int     cpus;
-    public final           long    megabytesOfRam;
+    public final           int    megabytesOfRam;
     public final @Nullable Integer gpus;
 
 
     @ConstructorProperties({"cpus", "megabytesOfRam", "gpus"})
-    public ResourceInfo(int cpus, long megabytesOfRam, @Nullable Integer gpus) {
+    public ResourceInfo(int cpus, int megabytesOfRam, @Nullable Integer gpus) {
         this.cpus           = cpus;
         this.megabytesOfRam = megabytesOfRam;
         this.gpus           = gpus;
